DDINVESTOR_COLLECTOR/data_collector/requirements.txt
pandas==1.0.3
Scrapy==2.0.1
selenium==3.141.0
scrapy-selenium==0.0.7
beautifulsoup4==4.8.2
tqdm==4.45.0
sqlalchemy==1.3.16
psycopg2-binary==2.8.5
yamldirs==1.1.8
lxml~=4.5.0
requests~=2.23.0
python-dateutil~=2.8.1
pyarrow~=0.17.0